Refactor responses into TestResponse
