+++ system/Test/ControllerResponse.php
@@ -11,33 +11,20 @@
 
 namespace CodeIgniter\Test;
 
-use CodeIgniter\HTTP\RedirectResponse;
-use CodeIgniter\HTTP\RequestInterface;
 use CodeIgniter\HTTP\ResponseInterface;
+use Config\Services;
 
 /**
  * Testable response from a controller
  */
-class ControllerResponse
+class ControllerResponse extends TestResponse
 {
-	/**
-	 * The request.
-	 *
-	 * @var RequestInterface
-	 */
-	protected $request;
-
-	/**
-	 * The response.
-	 *
-	 * @var ResponseInterface
-	 */
-	protected $response;
-
 	/**
 	 * The message payload.
 	 *
 	 * @var string
+	 *
+	 * @deprecated Use $response->getBody() instead
 	 */
 	protected $body;
 
@@ -45,35 +32,55 @@ class ControllerResponse
 	 * DOM for the body.
 	 *
 	 * @var DOMParser
+	 *
+	 * @deprecated Use $domParser instead
 	 */
 	protected $dom;
 
 	/**
-	 * Constructor.
+	 * Maintains the deprecated $domParser property.
 	 */
 	public function __construct()
 	{
-		$this->dom = new DOMParser();
+		parent::__construct($response ?? Services::response());
+
+		$this->dom = &$this->domParser;
 	}
 
-	//--------------------------------------------------------------------
-	// Getters / Setters
-	//--------------------------------------------------------------------
+	/**
+	 * Sets the response.
+	 *
+	 * @param ResponseInterface $response
+	 *
+	 * @return $this
+	 *
+	 * @deprecated Will revert to parent::setResponse() in a future release (no $body updates)
+	 */
+	public function setResponse(ResponseInterface $response)
+	{
+		parent::setResponse($response);
+
+		$this->body = $response->getBody() ?? '';
+
+		return $this;
+	}
 
 	/**
-	 * Set the body & DOM.
+	 * Sets the body and updates the DOM.
 	 *
 	 * @param string $body
 	 *
 	 * @return $this
+	 *
+	 * @deprecated Use response()->setBody() instead
 	 */
 	public function setBody(string $body)
 	{
 		$this->body = $body;
 
-		if (! empty($body))
+		if ($body !== '')
 		{
-			$this->dom = $this->dom->withString($body);
+			$this->domParser->withString($body);
 		}
 
 		return $this;
@@ -83,112 +90,11 @@ class ControllerResponse
 	 * Retrieve the body.
 	 *
 	 * @return string
+	 *
+	 * @deprecated Use response()->getBody() instead
 	 */
 	public function getBody()
 	{
 		return $this->body;
-	}
-
-	/**
-	 * Set the request.
-	 *
-	 * @param RequestInterface $request
-	 *
-	 * @return $this
-	 */
-	public function setRequest(RequestInterface $request)
-	{
-		$this->request = $request;
-
-		return $this;
-	}
-
-	/**
-	 * Set the response.
-	 *
-	 * @param ResponseInterface $response
-	 *
-	 * @return $this
-	 */
-	public function setResponse(ResponseInterface $response)
-	{
-		$this->response = $response;
-
-		$this->setBody($response->getBody() ?? '');
-
-		return $this;
-	}
-
-	/**
-	 * Request accessor.
-	 *
-	 * @return RequestInterface
-	 */
-	public function request()
-	{
-		return $this->request;
-	}
-
-	/**
-	 * Response accessor.
-	 *
-	 * @return ResponseInterface
-	 */
-	public function response()
-	{
-		return $this->response;
-	}
-
-	//--------------------------------------------------------------------
-	// Simple Response Checks
-	//--------------------------------------------------------------------
-
-	/**
-	 * Boils down the possible responses into a boolean valid/not-valid
-	 * response type.
-	 *
-	 * @return boolean
-	 */
-	public function isOK(): bool
-	{
-		// Only 200 and 300 range status codes
-		// are considered valid.
-		if ($this->response->getStatusCode() >= 400 || $this->response->getStatusCode() < 200)
-		{
-			return false;
-		}
-		// Empty bodies are not considered valid.
-		return ! empty($this->response->getBody());
-	}
-
-	/**
-	 * Returns whether or not the Response was a redirect or RedirectResponse
-	 *
-	 * @return boolean
-	 */
-	public function isRedirect(): bool
-	{
-		return $this->response instanceof RedirectResponse
-			|| $this->response->hasHeader('Location')
-			|| $this->response->hasHeader('Refresh');
-	}
-
-	//--------------------------------------------------------------------
-	// Utility
-	//--------------------------------------------------------------------
-
-	/**
-	 * Forward any unrecognized method calls to our DOMParser instance.
-	 *
-	 * @param  string $function Method name
-	 * @param  mixed  $params   Any method parameters
-	 * @return mixed
-	 */
-	public function __call($function, $params)
-	{
-		if (method_exists($this->dom, $function))
-		{
-			return $this->dom->{$function}(...$params);
-		}
 	}
 }
